requirements.txt
streamlit
numpy
matplotlib.pyplot
opencv-python-headless